54895/add import bulletin

See merge request 2021-2022/prjg5/equipe1!12

+++ database/migrations/2021_11_23_094653_create_demande_table.php
@@ -16,7 +16,7 @@ class CreateDemandeTable extends Migration
     {
         Schema::create('demande', function (Blueprint $table) {
             $table->id();
-            $table->integer('student_matricule');
+            $table->unsignedBigInteger('student_matricule');
             $table->foreign('student_matricule')->on('student')->references('matricule');
             $table->longText('message');
             $table->char('state', 1)->default('E');
